requirements.txt
python-telegram-bot==21.5
google-generativeai==0.8.3
openai==1.51.0
python-dotenv==1.0.1
